Add ACR server class
