+++ src/code/LocalServerApiCalls.cs
@@ -13,8 +13,9 @@ using System.Xml;
 using System.Net;
 using System.Management.Automation;
 using System.Runtime.ExceptionServices;
+using Microsoft.PowerShell.PSResourceGet.Cmdlets;
 
-namespace Microsoft.PowerShell.PSResourceGet.Cmdlets
+namespace Microsoft.PowerShell.PSResourceGet
 {
     internal class LocalServerAPICalls : ServerApiCall
     {
