+++ src/code/PSScriptFileInfo.cs
@@ -8,8 +8,9 @@ using System.IO;
 using System.Management.Automation;
 using System.Linq;
 using Microsoft.PowerShell.Commands;
+using Microsoft.PowerShell.PSResourceGet.UtilClasses;
 
-namespace Microsoft.PowerShell.PSResourceGet.UtilClasses
+namespace Microsoft.PowerShell.PSResourceGet
 {
     /// <summary>
     /// This class contains information for a PSScriptFileInfo (representing a .ps1 file contents).
