+++ src/code/PSScriptMetadata.cs
@@ -5,9 +5,10 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Management.Automation;
+using Microsoft.PowerShell.PSResourceGet.UtilClasses;
 using NuGet.Versioning;
 
-namespace Microsoft.PowerShell.PSResourceGet.UtilClasses
+namespace Microsoft.PowerShell.PSResourceGet
 {
     /// <summary>
     /// This class contains information for a PSScriptFileInfo (representing a .ps1 file contents).
