+++ src/code/PSScriptRequires.cs
@@ -9,7 +9,7 @@ using System.Linq;
 using System.Collections.ObjectModel;
 using Microsoft.PowerShell.Commands;
 
-namespace Microsoft.PowerShell.PSResourceGet.UtilClasses
+namespace Microsoft.PowerShell.PSResourceGet
 {
     /// <summary>
     /// This class contains information for a PSScriptFileInfo (representing a .ps1 file contents).
