+++ src/code/TestPSScriptFile.cs
@@ -4,9 +4,8 @@
 using System;
 using System.IO;
 using System.Management.Automation;
-using Microsoft.PowerShell.PSResourceGet.UtilClasses;
 
-namespace Microsoft.PowerShell.PSResourceGet.Cmdlets
+namespace Microsoft.PowerShell.PSResourceGet
 {
     /// <summary>
     /// Tests the contents of a .ps1 file to see if it has all properties and is in correct format
