+++ src/code/V2ServerAPICalls.cs
@@ -15,8 +15,9 @@ using System.Runtime.ExceptionServices;
 using System.Management.Automation;
 using System.Reflection;
 using System.Data.Common;
+using Microsoft.PowerShell.PSResourceGet.Cmdlets;
 
-namespace Microsoft.PowerShell.PSResourceGet.Cmdlets
+namespace Microsoft.PowerShell.PSResourceGet
 {
     internal class V2ServerAPICalls : ServerApiCall
     {
